+++ hahomematic/platforms/custom/light.py
@@ -17,7 +17,6 @@ from hahomematic.const import (
     HmEvent,
     HmPlatform,
 )
-from hahomematic.decorators import bind_collector
 from hahomematic.platforms import device as hmd
 from hahomematic.platforms.custom import definition as hmed
 from hahomematic.platforms.custom.const import (
@@ -44,12 +43,13 @@ from hahomematic.platforms.custom.const import (
 )
 from hahomematic.platforms.custom.entity import CustomEntity
 from hahomematic.platforms.custom.support import CustomConfig, ExtendedConfig
-from hahomematic.platforms.entity import CallParameterCollector
+from hahomematic.platforms.decorators import config_property, value_property
+from hahomematic.platforms.entity import CallParameterCollector, bind_collector
 from hahomematic.platforms.generic.action import HmAction
 from hahomematic.platforms.generic.number import HmFloat, HmInteger
 from hahomematic.platforms.generic.select import HmSelect
 from hahomematic.platforms.generic.sensor import HmSensor
-from hahomematic.platforms.support import OnTimeMixin, config_property, value_property
+from hahomematic.platforms.support import OnTimeMixin
 
 _HM_ARG_BRIGHTNESS: Final = "brightness"
 _HM_ARG_COLOR_NAME: Final = "color_name"