+++ hahomematic/platforms/custom/lock.py
@@ -9,7 +9,6 @@ from abc import abstractmethod
 from typing import Final
 
 from hahomematic.const import HmPlatform
-from hahomematic.decorators import bind_collector
 from hahomematic.platforms import device as hmd
 from hahomematic.platforms.custom import definition as hmed
 from hahomematic.platforms.custom.const import (
@@ -23,11 +22,11 @@ from hahomematic.platforms.custom.const import (
 )
 from hahomematic.platforms.custom.entity import CustomEntity
 from hahomematic.platforms.custom.support import CustomConfig, ExtendedConfig
-from hahomematic.platforms.entity import CallParameterCollector
+from hahomematic.platforms.decorators import value_property
+from hahomematic.platforms.entity import CallParameterCollector, bind_collector
 from hahomematic.platforms.generic.action import HmAction
 from hahomematic.platforms.generic.sensor import HmSensor
 from hahomematic.platforms.generic.switch import HmSwitch
-from hahomematic.platforms.support import value_property
 
 # HM constants
 _LOCK_STATE_UNKNOWN: Final = "UNKNOWN"