+++ hahomematic/platforms/custom/siren.py
@@ -9,7 +9,6 @@ from abc import abstractmethod
 from typing import Any, Final, TypedDict
 
 from hahomematic.const import HmPlatform
-from hahomematic.decorators import bind_collector
 from hahomematic.platforms import device as hmd
 from hahomematic.platforms.custom import definition as hmed
 from hahomematic.platforms.custom.const import (
@@ -25,11 +24,11 @@ from hahomematic.platforms.custom.const import (
 )
 from hahomematic.platforms.custom.entity import CustomEntity
 from hahomematic.platforms.custom.support import CustomConfig, ExtendedConfig
-from hahomematic.platforms.entity import CallParameterCollector
+from hahomematic.platforms.decorators import config_property, value_property
+from hahomematic.platforms.entity import CallParameterCollector, bind_collector
 from hahomematic.platforms.generic.action import HmAction
 from hahomematic.platforms.generic.binary_sensor import HmBinarySensor
 from hahomematic.platforms.generic.sensor import HmSensor
-from hahomematic.platforms.support import config_property, value_property
 
 _HM_ARG_ACOUSTIC_ALARM: Final = "acoustic_alarm"
 _HM_ARG_OPTICAL_ALARM: Final = "optical_alarm"