+++ hahomematic/support.py
@@ -1,13 +1,12 @@
 """Helper functions used within hahomematic."""
 from __future__ import annotations
 
-import asyncio
 import base64
 from collections.abc import Callable, Collection
 import contextlib
 from dataclasses import dataclass, field
 from datetime import datetime
-from functools import cache, wraps
+from functools import cache
 import logging
 import os
 import re
@@ -30,7 +29,7 @@ from hahomematic.const import (
 )
 from hahomematic.exceptions import HaHomematicException
 
-_LOGGER = logging.getLogger(__name__)
+_LOGGER = logging.getLogger("hahomematic.base")
 
 _CallableT = TypeVar("_CallableT", bound=Callable[..., Any])
 
@@ -301,49 +300,3 @@ class Channel:
     def no(self) -> int | None:
         """Return the channel no."""
         return get_channel_no(self.address)
-
-
-def measure_execution_time(func: _CallableT) -> _CallableT:
-    """Decorate function to measure the function execution time."""
-
-    is_enabled = _LOGGER.isEnabledFor(level=logging.DEBUG)
-
-    @wraps(func)
-    async def async_wrapper(*args: Any, **kwargs: Any) -> Any:
-        """Wrap method."""
-        if is_enabled:
-            start = datetime.now()
-        try:
-            return await func(*args, **kwargs)
-        finally:
-            if is_enabled:
-                delta = (datetime.now() - start).total_seconds()
-                _LOGGER.info(
-                    "Execution of %s took %ss args(%s) kwargs(%s) ",
-                    func.__name__,
-                    delta,
-                    args,
-                    kwargs,
-                )
-
-    @wraps(func)
-    def wrapper(*args: Any, **kwargs: Any) -> Any:
-        """Wrap method."""
-        if is_enabled:
-            start = datetime.now()
-        try:
-            return func(*args, **kwargs)
-        finally:
-            if is_enabled:
-                delta = (datetime.now() - start).total_seconds()
-                _LOGGER.info(
-                    "Execution of %s took %ss args(%s) kwargs(%s) ",
-                    func.__name__,
-                    delta,
-                    args,
-                    kwargs,
-                )
-
-    if asyncio.iscoroutinefunction(func):
-        return async_wrapper  # type: ignore[return-value]
-    return wrapper  # type: ignore[return-value]