+++ hahomematic/xml_rpc_proxy.py
@@ -15,7 +15,7 @@ from hahomematic import central_unit as hmcu
 from hahomematic.exceptions import AuthFailure, ClientException, NoConnection
 from hahomematic.support import get_tls_context, reduce_args
 
-_LOGGER = logging.getLogger(__name__)
+_LOGGER = logging.getLogger("hahomematic.client")
 
 _T = TypeVar("_T")